hw04/requirements.txt
addict==2.4.0
albucore==0.0.17
albumentations==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
collection==0.1.6
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
einops==0.8.1
ensemble-boxes==1.0.9
eval_type_backport==0.2.2
filelock==3.14.0
fonttools==4.57.0
fsspec==2025.3.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.41.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmcv==2.1.0
-e git+https://github.com/open-mmlab/mmdetection.git@cfd5d3a985b0249de009b67d04f37263e11cdf3d#egg=mmdet
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.2
qudida==0.0.4
requests==2.28.2
rich==13.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
tomli==2.2.1
torch==1.8.1+cu101
torchaudio==0.8.1
torchvision==0.9.1+cu101
tqdm==4.65.2
triton==3.0.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==1.26.20
yapf==0.43.0
zipp==3.20.2
